# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Red Lycoris API",
    "version": "1.0.0",
    "description": "Stable API contracts for v1. Responses are additive-only (new optional fields). Request bodies ignore unknown fields; clients must avoid unknown fields on PUT replace operations."
  },
  "paths": {
...
    "/api/v1/gates/check": {
      "post": {
        "summary": "Gate check for an import job",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GateCheckRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Gate check response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GateCheckResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "Import job not found"
          },
          "412": {
            "description": "Gate failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GateCheckResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "GateCheckRequest": {
        "type": "object",
        "required": [
          "importJobId"
        ],
        "properties": {
          "importJobId": {
            "type": "string",
            "format": "uuid"
          },
          "productId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "profile": {
            "type": "string",
            "nullable": true
          },
          "commitSha": {
            "type": "string",
            "nullable": true
          },
          "buildId": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "GateCheckPolicy": {
        "type": "object",
        "properties": {
          "policyId": {
            "type": "string",
            "nullable": true
          },
          "policyRuleId": {
            "type": "string",
            "nullable": true
          },
          "version": {
            "type": "string",
            "nullable": true
          },
          "sha256": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "GateBlockingFinding": {
        "type": "object",
        "properties": {
          "findingId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "severity": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "violationCode": {
            "type": "string"
          }
        }
      },
      "GateCheckViolation": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "severity": {
            "type": "string"
          },
          "refs": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "GateCheckResponse": {
        "type": "object",
        "properties": {
          "pass": {
            "type": "boolean"
          },
          "decision": {
            "type": "string",
            "enum": [
              "pass",
              "warn",
              "fail"
            ]
          },
          "blockingFindings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GateBlockingFinding"
            }
          },
          "violations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GateCheckViolation"
            }
          },
          "policy": {
            "$ref": "#/components/schemas/GateCheckPolicy"
          },
          "evaluatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}